requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.5.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2024.8.30; python_version >= '3.6'
distro==1.9.0; python_version >= '3.6'
django-cors-headers==4.4.0
django==4.2.16
djangorestframework==3.15.2
exceptiongroup==1.2.2; python_version < '3.11'
gunicorn==23.0.0
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.5.0; python_version >= '3.8'
numpy==1.24.4; python_version < '3.10'
openai==1.50.2
packaging==24.1; python_version >= '3.8'
pandas==2.0.3
pydantic-core==2.23.4; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
python-env==1.0.0
pytz==2024.2
six==1.16.0
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.1; python_version >= '3.8'
tqdm==4.66.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
